+++ doc/whats_new.rst
@@ -55,6 +55,8 @@ Enhancements
   to automatically add an attributes section to the docstring of the class to
   which the decorator is applied, using the metadata of each variable declared
   in the class (:issue:`67`).
+- Added :func:`~xsimlab.validators.in_bounds` and
+  :func:`~xsimlab.validators.is_subdtype` validators (:issue:`87`).
 
 Bug fixes
 ~~~~~~~~~